Add a dedicated message queue and break input
I'm not sure if this is the best way that's most consistent with the
rest of the engine. I tried to vaguely follow the lines of `Input` and
most importantly `MessageQueue`, but I might as well have made
interpreted them in the wrong way, focusing for now on making it work
before anything else.

I still can't test easily if the `call_deferred` issue persists because
`Input` dispatching is currently `#if 0`'ed out, but I think that it
should work(tm) in that case, or so I hope.

For now I've just added the simplest queue possible, with one single
"base" class, containing an enum that contains a void pointer to some
other data defined by an enum.

The data must be allocated on event pushing in the queue and
deallocated on popping from the queue.

Also, I changed some minor things in the window data, like removing the
now unused Vulkan context handle (which IMO shouldn't even be there but
as I said there's still time to polish things up and things must work
first before doing that).

And yes, I know, WaylandMessage is a pretty bad name. I hope I'll find a
better one sometime, but for now it isn't of absolute importance.