Add `InputEvent`s to the message queue, fix Input
Now we're back again to the old state of the project but without
Wayland's limited event buffer nor weird conflicts with MessageQueue.

I feel like I must note that `Input` seems to already support some sort
of buffering with subsequent release on `process_events` and thus in
the main thread, but at a first glance it looks like it can get
disabled, so at least for now I'm adding it to our own queue which is
garaunteed to always buffer incoming events until the next call to
`process_events`.