More thorough heuristic and actual asynchronous destruction
I think I found a project setting to disable resizing, so let's just
add that the window has to be also borderless to be considered a popup.

We could additionally filter the main window out by forcing a popup to
be initialized with a parent (the main window can't have a parent) but
I'm lazy and this is all temporary anyways - this heuristic is just a
quick hack to keep everything toghether while I finish the foundations
for multiwin.

Still gotta handle the events which allow "null" surfaces such as the
keyboard exit one. I'll have to pass through all of the "allow null"
wl_surface-related events eventually, if there's more.

It looks like those are the only kind of meaningful events left after
destroying a window's objects so we can roundtrip the wl_display after
deleting everything instead of doing it half-way in the destruction.

Also, for some reason, while everything is "fine" on sway, input
handling is very wonky on Weston, almost as if we passed the position to
the child toplevel, although we certainly do not.