What if... We moved the deletion stuff in the server (+ other fixes)
I still have to figure out all the cases but the only thing I'm certain
about is that we _have_ to add some certain callback thing that asks for
a full, proper destruction.

Certain window destruction things _must_ be done by the compositor,
famously the popup stack thing, which is already a bit ambiguous on its
own. Add to that the fact that there's a binding in the renderer done
from the respective node and one done in the DisplayServer and you can
tell that there's no pre-existing way of satisfying our requirements.

I'll look into a solution another time because I'm tired.

Oh, also, the popup stack variable is currently global but should be per
toplevel. I'll definitely not forget that.